slate sub-foldering added

+++ 5_app/sort_by_slate.py
@@ -1,7 +1,7 @@
 '''*************************************************
 content     Tool Demo
 
-version     0.0.1
+version     0.0.2
 date        14-12-2025
 
 author      Harry Shaper <harryshaper@gmail.com>
@@ -57,8 +57,10 @@ def get_shoot_data_types():
 def update_slate_list(folder):
     data_set_folder = os.listdir(folder)
     for item in data_set_folder:
-        if item not in SLATE_LIST:
-            SLATE_LIST.append(item)
+        unique_slate = item.split("_")
+        if unique_slate[0] not in SLATE_LIST:
+            SLATE_LIST.append(unique_slate[0])
+    
 
 #LOOK THROUGH ALL SHOOT DATA
 def get_slates():
@@ -66,8 +68,6 @@ def get_slates():
         file_path = SHOOT_FOLDER + "\\" + item
         #print(file_path)
         update_slate_list(file_path)
-
-    #print(SLATE_LIST)
 
 #Makes all necessary slate folder
 def make_slate_folders():
@@ -86,30 +86,40 @@ def sort_data():
     moves = []
     source_folders = set(ORIGINAL_DATA_FOLDERS)
 
-    #Scan through items
-    for folder in define_shoot_data():
-        current_data_type = os.path.basename(folder)
-        source_folders.add(folder)
+    # Scan through each shoot data type (HDRI / PANO / TEXTURE / etc)
+    for data_type_folder in define_shoot_data():
+        data_type = os.path.basename(data_type_folder)
 
-        for sub_folder in os.listdir(folder):
-            current_path = os.path.join(folder, sub_folder)
-            destination = os.path.join(
+        for item in os.listdir(data_type_folder):
+            src_path = os.path.join(data_type_folder, item)
+
+            # --- CHANGE: ensure we only process folders ---
+            if not os.path.isdir(src_path):
+                continue
+
+            slate = item.split("_")[0].upper()
+
+            # --- CHANGE: destination now uses slate, not full folder name ---
+            dst_path = os.path.join(
                 SHOOT_FOLDER,
-                sub_folder,
-                current_data_type
+                slate,
+                data_type,
+                item  
             )
 
-            moves.append((current_path, destination))
+            moves.append((src_path, dst_path))
 
-    #Move items
+    # Move items
     for src, dst in moves:
-        os.makedirs(dst, exist_ok=True)
+        # --- CHANGE: only create parent directories ---
+        os.makedirs(os.path.dirname(dst), exist_ok=True)
         shutil.move(src, dst)
 
-    #Clear original folders
+    # Clear original folders
     for folder in source_folders:
-        if not os.listdir(folder):
+        if os.path.exists(folder) and not os.listdir(folder):
             os.rmdir(folder)
+
         
 
 get_slates()
